requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Dataset and model hub
datasets>=2.12.0
huggingface_hub>=0.15.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.8.0  # For camera streaming and real-time detection

# Scientific computing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0

# Progress bars and utilities
tqdm>=4.65.0

# Jupyter notebook support
ipywidgets>=8.0.0
notebook>=6.5.0

# Data augmentation
albumentations>=1.3.0

# YAML configuration
PyYAML>=6.0

# Validation and type checking
pydantic>=1.10.0

# Logging
loguru>=0.7.0

# System utilities
psutil>=5.9.0

# Additional computer vision
scikit-image>=0.20.0

# Model optimization
onnx>=1.14.0
onnxruntime>=1.15.0

# Async support
aiofiles>=23.0.0

# Configuration management
hydra-core>=1.3.0

# Experiment tracking
wandb>=0.15.0

# Web interface (optional)
gradio>=3.35.0
streamlit>=1.24.0

# HTTP requests
requests>=2.31.0

# Audio processing (if needed)
librosa>=0.10.0

# Advanced image processing
opencv-contrib-python>=4.8.0  # Additional OpenCV modules

# Performance monitoring
memory-profiler>=0.60.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Documentation
sphinx>=7.0.0

# Development utilities
ipython>=8.14.0

# Deployment
fastapi>=0.100.0
uvicorn>=0.23.0

# Model conversion
coremltools>=6.3.0
tensorflow>=2.13.0  # For some conversion utilities
